feat: Add amount property to CryptoConv for original currency conversion
- Implemented an `amount` property in the `CryptoConv` class to return the conversion value in the original currency as an `Amount` object.
- Updated the `process_payment_success` function to handle payment processing with Keepsats, including conversion from Hive if necessary.
- Introduced a new `process_hive.py` module to handle various Hive transaction operations, including transfers and order processing.
- Enhanced error handling for ledger entry creation and processing of custom JSON operations.
- Adjusted the maximum lines per page in the balance sheet RTF generation from 50 to 45 for better formatting.
